[MTDSA-434] Refactor liability calculator
Each calculation step can return a liability or an error which is
persisted and subsequently retrieved from its endpoint.
The liability calculator fails fast on the occurrence of an error.

Remove the verification of the expected Json body from the the liability
functional tests.  The Json structure is done at the unit level when serializing the
liability response.